+++ base/time.c
@@ -1,4 +1,4 @@
-inline fn bool is_leap_year(u32 year) {
+inline fn bool is_leap_year(i32 year) {
     return (year % 4 == 0 && year % 100 != 0) ||
            (year % 400 == 0);
 }
@@ -11,7 +11,7 @@ fn DateTime datetime_from_time64(time64 t) {
   res.hour   = (u8)((t >> 22) & bitmask5);
   res.day    = (u8)((t >> 27) & bitmask5);
   res.month  = (u8)((t >> 32) & bitmask4);
-  res.year   = ((t >> 36) & ~bit28) * (t >> 63 ? 1 : -1);
+  res.year   = (i32)((t >> 36) & ~bit28) * (t >> 63 ? 1 : -1);
   return res;
 }
 
@@ -58,7 +58,7 @@ fn DateTime datetime_from_unix(u64 timestamp) {
 
 fn time64 time64_from_datetime(DateTime *dt) {
   time64 res = (dt->year >= 0 ? 1ULL << 63 : 0);
-  res |= (u64)((dt->year >= 0 ? dt->year : -dt->year) & ~bit28) << 36;
+  res |= ((u64)(dt->year >= 0 ? dt->year : -dt->year) & ~bit28) << 36;
   res |= (u64)(dt->month) << 32;
   res |= (u64)(dt->day) << 27;
   res |= (u64)(dt->hour) << 22;
@@ -71,7 +71,7 @@ fn time64 time64_from_datetime(DateTime *dt) {
 fn time64 time64_from_unix(u64 timestamp) {
   time64 res = 1ULL << 63;
 
-  u32 year = 1970;
+  i32 year = 1970;
   for (u64 secondsXyear = is_leap_year(year)
                           ? UNIX_LEAP_YEAR
                           : UNIX_YEAR;
@@ -81,7 +81,7 @@ fn time64 time64_from_unix(u64 timestamp) {
                               ? UNIX_LEAP_YEAR
                               : UNIX_YEAR);
 
-  res |= ((u64)year & ~(1 << 27)) << 36;
+  res |= (u64)((i64)year & ~(1 << 27)) << 36;
   u64 month = 1;
   while (1) {
     u8 days = daysXmonth[month - 1];
@@ -120,7 +120,7 @@ fn u64 unix_from_datetime(DateTime *dt) {
                   (dt->minute * UNIX_MINUTE) +
                   (dt->second);
 
-  for (u32 year = 1970; year < (u32)dt->year; ++year) {
+  for (i32 year = 1970; year < (i32)dt->year; ++year) {
     unix_time += is_leap_year(year) ? UNIX_LEAP_YEAR : UNIX_YEAR;
   }
 
@@ -138,7 +138,7 @@ fn u64 unix_from_time64(time64 timestamp) {
   u64 res = 0;
   if (!(timestamp >> 63)) { return 0; }
 
-  i32 year = ((timestamp >> 36) & ~(1 << 27));
+  i32 year = ((i32)(timestamp >> 36) & ~(1 << 27));
   if (year < 1970) { return 0; }
   for (i32 i = 1970; i < year; ++i) {
     res += is_leap_year(i) ? UNIX_LEAP_YEAR : UNIX_YEAR;
